@@ -13,21 +13,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.0.2",
-    "@angular/common": "17.0.2",
-    "@angular/compiler": "17.0.2",
-    "@angular/core": "17.0.2",
-    "@angular/forms": "17.0.2",
-    "@angular/platform-browser": "17.0.2",
-    "@angular/platform-browser-dynamic": "17.0.2",
-    "@angular/router": "17.0.2",
+    "@angular/animations": "17.3.6",
+    "@angular/common": "17.3.6",
+    "@angular/compiler": "17.3.6",
+    "@angular/core": "17.3.6",
+    "@angular/forms": "17.3.6",
+    "@angular/platform-browser": "17.3.6",
+    "@angular/platform-browser-dynamic": "17.3.6",
+    "@angular/router": "17.3.6",
     "@capacitor-community/http": "1.4.1",
     "@capacitor/app": "6.0.0",
     "@capacitor/core": "6.0.0",
+    "@capacitor/device": "^6.0.0",
     "@capacitor/haptics": "6.0.0",
     "@capacitor/keyboard": "6.0.0",
     "@capacitor/status-bar": "6.0.0",
     "@ionic/angular": "8.0.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "crypto-js": "4.2.0",
     "ionicons": "7.2.1",
     "rxjs": "~7.8.0",
@@ -35,15 +38,15 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.0.0",
+    "@angular-devkit/build-angular": "17.3.6",
     "@angular-eslint/builder": "17.0.0",
     "@angular-eslint/eslint-plugin": "17.0.0",
     "@angular-eslint/eslint-plugin-template": "17.0.0",
     "@angular-eslint/schematics": "17.0.0",
     "@angular-eslint/template-parser": "17.0.0",
-    "@angular/cli": "17.0.0",
-    "@angular/compiler-cli": "17.0.2",
-    "@angular/language-service": "17.0.2",
+    "@angular/cli": "17.3.6",
+    "@angular/compiler-cli": "17.3.6",
+    "@angular/language-service": "17.3.6",
     "@capacitor/cli": "6.0.0",
     "@ionic/angular-toolkit": "11.0.1",
     "@types/crypto-js": "^4.2.2",
